+++ src/youwol/app/environment/models/models_features.py
@@ -1,0 +1,9 @@
+# third parties
+from pydantic import BaseModel
+
+
+class Features(BaseModel):
+    """Application Features"""
+
+    configDependantBrowserCaching: bool = False
+    """ Tie the browser caching to the configuration """