+++ src/youwol/app/middlewares/browser_caching_middleware.py
@@ -10,6 +10,12 @@ from starlette.requests import Request
 from starlette.responses import Response
 from starlette.types import ASGIApp
 
+# Youwol utilities
+from youwol.utils.crypto.digest import compute_digest
+
+# relative
+from ..environment.youwol_environment import YouwolEnvironmentFactory
+
 
 class BrowserCachingMiddleware(BaseHTTPMiddleware):
     """
@@ -19,16 +25,22 @@ class BrowserCachingMiddleware(BaseHTTPMiddleware):
     cache = {}
 
     def __init__(
-        self, app: ASGIApp, dispatch: DispatchFunction | None = None, **_
+        self,
+        app: ASGIApp,
+        dispatch: DispatchFunction | None = None,
+        config_dependant_browser_caching=False,
+        **_,
     ) -> None:
         """
         Initializes a new instance of BrowserCachingMiddleware.
 
         Parameters:
            app: The ASGI application, forwarded to `BaseHTTPMiddleware`.
+           config_dependant_browser_caching: Enable cache dependent_browser_cache
            dispatch: The dispatch function, forwarded to `BaseHTTPMiddleware`.
         """
         super().__init__(app, dispatch)
+        self.config_dependant_browser_caching = config_dependant_browser_caching
 
     async def dispatch(
         self, request: Request, call_next: RequestResponseEndpoint
@@ -46,6 +58,25 @@ class BrowserCachingMiddleware(BaseHTTPMiddleware):
             The response with eventually modified headers.
         """
         response = await call_next(request)
+        if self.config_dependant_browser_caching:
+            digest = compute_digest(
+                {
+                    "config_digest": YouwolEnvironmentFactory.get_digest(),
+                    "user_info": (
+                        {
+                            "name": request.state.user_info["sub"],
+                            "groups": request.state.user_info["memberof"],
+                        }
+                        if hasattr(request.state, "user_info")
+                        else None
+                    ),
+                },
+                trace_path_root="browser_caching_middleware",
+            ).hex()
+            response.headers["Youwol-Config-Digest"] = digest
+            response.set_cookie("youwol-config-digest", digest)
+            response.headers["Vary"] = "Youwol-Config-Digest, Cookie"
+            response.headers["Cache-Control"] = "max-age=3600"
         if "/api/assets-gateway/raw/" in request.url.path and request.method == "GET":
             response.headers["Cache-Control"] = "no-cache, no-store"